requirements.txt
torch
torchvision
transformers
rembg
trimesh
omegaconf==2.3.0
einops==0.7.0
PyMCubes
huggingface-hub
imageio[ffmpeg]
moderngl==5.10.0
xatlas
fastapi
uvicorn
python-multipart
numpy
Pillow
gradio
